requirements.txt
Click==7.0
dnspython==1.16.0
eventlet==0.24.1
Flask==0.12.2
Flask-SocketIO==2.9.6
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
monotonic==1.5
Pillow==5.4.1
python-engineio==3.3.0
python-socketio==3.1.1
six==1.12.0
Werkzeug==0.14.1
